Support `as` expression in `#const` macro.

+++ Tests/SQLGrammarTests/SQLGrammarTests.swift
@@ -2219,7 +2219,7 @@ final class SQLGrammarMacroExpansionTests: XCTestCase {
   func test_const() {
     XCTAssertEqual(#const("My String").description, #"'My String'"#)
     XCTAssertEqual(#const(12345).description, #"12345"#)
-    XCTAssertEqual(#const(18446744073709551615).description, "18446744073709551615")
+    XCTAssertEqual(#const(18446744073709551615 as UInt64).description, "18446744073709551615")
     XCTAssertEqual(#const(123.45).description, #"123.45"#)
     XCTAssertEqual(#const(+12345).description, #"+12345"#)
     XCTAssertEqual(#const(+123.45).description, #"+123.45"#)
